+++ NEWS.md
@@ -39,6 +39,7 @@ cd lib/sql; sudo make all install
 * Properly and fully disallowing multilogins, by killing other existing sessions upon login
 * Added new Platform Limit for concurrent syncs per user. Currently 2 hours TTL, 3 syncs per user max. allowed
 * Importer now tries to import shapefile zips without .prj file, by setting 4326 projection.
+* Synchronizations model now has a new field (and FK) to visualizations.id and joins to them using that instead of by matching name to canonical visualization's table name. It also gets deleted if FK dissapears.
 
 Bugfixes:
 * Fixed deletion of layers upon disconnecting synced datasources [#3718](https://github.com/CartoDB/cartodb/pull/3718)